+++ Sources/RoktUXHelper/Services/ActionCollection.swift
@@ -11,13 +11,27 @@
 
 import Foundation
 
+/// Represents the various types of user or system actions that can occur within the Rokt experience.
 enum RoktActionType: String {
     case close
     case nextOffer
     case unload
     case positiveEngaged
-    case nextGroup
-    case previousGroup
+
+    /// Informs listening distributions to navigate to next. Next will be different depending on the distrution e.g.
+    /// ``OneByOneDistributionComponent`` - would go to next offer
+    /// ``GroupedDistributionComponent`` - would go to the next page or group
+    ///
+    /// Triggered by ``ProgressControlComponent``
+    case progressControlNext
+
+    /// Informs listening distributions to navigate to previous. Previous will be different depending on the distrution e.g.
+    /// ``OneByOneDistributionComponent`` - would go to previous offer
+    /// ``GroupedDistributionComponent`` - would go to the previous page or group
+    ///
+    /// Triggered by ``ProgressControlComponent``
+    case progressControlPrevious
+
     case checkBoundingBox
     case checkBoundingBoxMissized
     case toggleCustomState