+++ Sources/RoktUXHelper/UI/Components/Distribution/OneByOneDistributionComponent.swift
@@ -162,6 +162,8 @@ struct OneByOneDistributionComponent: View {
     }
 
     func registerActions() {
+        model.layoutState?.actionCollection[.progressControlPrevious] = goToPreviousOffer
+        model.layoutState?.actionCollection[.progressControlNext] = goToNextOffer
         model.layoutState?.actionCollection[.nextOffer] = goToNextOffer
         model.layoutState?.actionCollection[.toggleCustomState] = toggleCustomState
         model.setupBindings(
@@ -183,6 +185,12 @@ struct OneByOneDistributionComponent: View {
             }
 
             exit()
+        }
+    }
+
+    func goToPreviousOffer(_: Any? = nil) {
+        if currentOffer != 0 {
+            transitionToPreviousOffer()
         }
     }
 
@@ -221,6 +229,26 @@ struct OneByOneDistributionComponent: View {
         default:
             self.customStateMap = RoktUXCustomStateMap()
             self.currentOffer = currentOffer + 1
+        }
+    }
+
+    func transitionToPreviousOffer() {
+        switch transition {
+        case .fadeInOut(let settings):
+            let duration = Double(settings.duration)/1000/2
+            withAnimation(.easeOut(duration: duration)) {
+                toggleTransition = false
+            }
+
+            // Wait to complete fade out of previous offer
+            // Must not run on `main` as that prevents `@State` from changing
+            DispatchQueue.background.asyncAfter(deadline: .now() + duration) {
+                self.customStateMap = RoktUXCustomStateMap()
+                self.currentOffer = currentOffer - 1
+            }
+        default:
+            self.customStateMap = RoktUXCustomStateMap()
+            self.currentOffer = currentOffer - 1
         }
     }