+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add support to all distributions for clickable progress
+
 ### Fixed
 
 - Ignore sub-pixel height changes by rounding before size-change callback